Update README.md
I struggle for a while figure out how to build this addon for LV2021 and found that the config files were forcing LabVIEW to look at old DLL assemblies.  So, I wanted to add this note to the readme in hopes that anyone in the future (including myself) knows to try that if they are having issues getting it to build.